Move VariantType to VariantType.py
Less circular dependencies for imports.
